app/backend/requirements.txt
#### Any version change made here should also be made and tested for the enrichment and function apps in /functions and /app/enrichment
azure-cosmos==4.7.0
azure-identity==1.18.0
azure-mgmt-cognitiveservices==13.5.0
azure-search-documents==11.5.1
azure-storage-blob==12.23.0
cryptography==43.0.3
fastapi==0.115.0
fastapi-utils==0.7.0
Flask==3.0.3
gunicorn==23.0.0
langchain==0.3.9
langchain-community==0.3.1 
langchain-experimental==0.3.2
langchain-openai==0.2.1
matplotlib==3.9.2
microsoft-bing-websearch==1.0.0
numexpr==2.10.1
openai==1.55.3
pandas==2.2.3
Pillow==10.4.0
PyJWT==2.10.1
pytest==8.3.3
python-dotenv==1.0.1
python-multipart==0.0.18
tabulate==0.9.0
tiktoken==0.7.0
uvicorn==0.30.6
wikipedia==1.4.0
pydantic==2.10.1
rich==13.9.4